requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-cosmos==4.9.0
azure-storage-blob==12.24.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
isodate==0.7.2
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
